fix: 공유하기 안내 모달로 변경 #144
